Move non-electron main utils out of utils.ts
utils.ts is a very poorly named library, as it primarily exists
for various bridge calls between the renderer and the main process.
Whereas, we actually would like some utility functions that have
relatively few (if any) dependencies, so that we may test them
better with a node runtime.

This is necessary for us to use mocha for newer NPMs, as build
dependencies are not shaken consistently across newer NPMs.
